Began implementing repeat functionality.
Repeat has been added as a function that should either report or set a repeat mode. However, it seems that my understanding of stream players was off a bit; a player may only play once, so if I want to repeat a song, I will need to construct a new player for it.

This is making me reconsider making the jump to more intelligently planned queueing sooner.