Converted Song_Queue to Music_Class, Rewrote Play Execute
Since the song_queue class was handling playback, queueing, and caching, I decided to rename it to better describe its functionality.

Play.py's execute function had some nasty formatting for determining whether or not to queue a song or play one, so I rewrote it, fixing a bug with queueing songs while a song was paused along the way.

I think there are some other minor bug fixes in here too (at least one in how the bot reports queued songs when readding songs to a queue in the two repeat modes).