Working on allowing playlist queueing
Added some debug text to information extraction phase of loading songs from urls to determine that yes, in fact, youtube-dl supports extracting playlists.

Split custom_create_ytdl_player into two functions - one that extracts information about a link and returns it, and another that takes the extracted information and creates a player with it. This is laying the groundwork of extracting and caching the songs from a playlist in the first half of the process, which will then be returned as a list of songs that will each individually be added with the second function.