Reading in permissions from config file and applying to bot by user role
Permissions by role are being read in from permissions.ini.

User permissions are then set based on the roles and role permissions of a user querying the bot. As this sets the current permission level of the bot as a whole instead of the user or message, this feels a bit hacky, but I'm not able to add attributes to users or messages at present, so this will have to do for now. I will likely re-evaluate this in the future.