Commands restricted by permission level, minor restructuring
Commands are now restricted by user roles and permissions.

Additionally switched a number of messages over to the debug logger to keep them from cluttering normal log output.

Added an additional role to the permissions.ini file; a user with the BB-DJ role will have access to all music related functions. Also added skip to the default role, as I want to later add voting to the skip command and they will need access to the command to use it.